requirements.txt
pyTelegramBotAPI
requests
pandas
python-dotenv
psycopg2